yahoo home page launched!!!!

+++ ECMS.WebV2/AppCode/HtmlExtension.cs
@@ -15,7 +15,7 @@ namespace ECMS.WebV2.AppCode
     {
         public static MvcHtmlString SSI(this HtmlHelper htmlHelper, string partialViewName)
         {
-            var viewName = string.Format("~/Views/{0}/{1}/{2}.cshtml", ECMSSettings.Current.SiteId, Utility.CurrentViewType(new HttpContextWrapper(HttpContext.Current)), partialViewName);
+            var viewName = string.Format("~/Views/{0}/{1}/{2}.cshtml", ECMSSettings.Current.SiteId, (int)Utility.CurrentViewType(new HttpContextWrapper(HttpContext.Current)), partialViewName);
             return htmlHelper.Partial(viewName);
         }
     }
